pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.124.4",
    "playwright>=1.57.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
]
